requirements.txt
dj-database-url==0.4.1
Django==1.9.7
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==3.2
djangorestframework==3.4.0
django-rest-auth==0.7.0
django-allauth==0.25.2
python-openid==2.2.5
oauthlib==1.1.2

